added custom hook for data fetching of Mtg Inventory
